refactor(command): parse-command is now a separate module
As parseCommand() is used by argsert, completion, etc.
