+++ lib/argsert.js
@@ -2,7 +2,7 @@
 
 // hoisted due to circular dependency on command.
 module.exports = argsert
-const command = require('./command')()
+const { parseCommand } = require('../build/lib/parse-command')
 const { YError } = require('../build/lib/yerror')
 
 const positionName = ['first', 'second', 'third', 'fourth', 'fifth', 'sixth']
@@ -17,7 +17,7 @@ function argsert (expected, callerArguments, length) {
       length = callerArguments
       callerArguments = expected
     } else {
-      parsed = command.parseCommand(`cmd ${expected}`)
+      parsed = parseCommand(`cmd ${expected}`)
     }
     const args = [].slice.call(callerArguments)
