+++ lib/command.js
@@ -3,6 +3,7 @@
 const inspect = require('util').inspect
 const { isPromise } = require('../build/lib/is-promise')
 const { applyMiddleware, commandMiddlewareFactory } = require('./middleware')
+const { parseCommand } = require('../build/lib/parse-command')
 const path = require('path')
 const Parser = require('yargs-parser')
 
@@ -40,10 +41,10 @@ module.exports = function command (yargs, usage, validation, globalMiddleware) {
     }
 
     // parse positionals out of cmd string
-    const parsedCommand = self.parseCommand(cmd)
+    const parsedCommand = parseCommand(cmd)
 
     // remove positional args from aliases only
-    aliases = aliases.map(alias => self.parseCommand(alias).cmd)
+    aliases = aliases.map(alias => parseCommand(alias).cmd)
 
     // check for default and filter out '*''
     let isDefault = false
@@ -131,34 +132,6 @@ module.exports = function command (yargs, usage, validation, globalMiddleware) {
       if (typeof test === 'string' || typeof test === 'boolean') return test
     }
     return false
-  }
-
-  self.parseCommand = function parseCommand (cmd) {
-    const extraSpacesStrippedCommand = cmd.replace(/\s{2,}/g, ' ')
-    const splitCommand = extraSpacesStrippedCommand.split(/\s+(?![^[]*]|[^<]*>)/)
-    const bregex = /\.*[\][<>]/g
-    const parsedCommand = {
-      cmd: (splitCommand.shift()).replace(bregex, ''),
-      demanded: [],
-      optional: []
-    }
-    splitCommand.forEach((cmd, i) => {
-      let variadic = false
-      cmd = cmd.replace(/\s/g, '')
-      if (/\.+[\]>]/.test(cmd) && i === splitCommand.length - 1) variadic = true
-      if (/^\[/.test(cmd)) {
-        parsedCommand.optional.push({
-          cmd: cmd.replace(bregex, '').split('|'),
-          variadic
-        })
-      } else {
-        parsedCommand.demanded.push({
-          cmd: cmd.replace(bregex, '').split('|'),
-          variadic
-        })
-      }
-    })
-    return parsedCommand
   }
 
   self.getCommands = () => Object.keys(handlers).concat(Object.keys(aliasMap))
@@ -403,7 +376,7 @@ module.exports = function command (yargs, usage, validation, globalMiddleware) {
       demand: {}
     }
 
-    const parsed = self.parseCommand(cmdString)
+    const parsed = parseCommand(cmdString)
     parsed.demanded.forEach((d) => {
       const cmds = d.cmd.slice(0)
       const cmd = cmds.shift()
