+++ lib/parse-command.ts
@@ -1,0 +1,30 @@
+import { ParsedCommand } from './types'
+
+export function parseCommand (cmd: string) {
+  const extraSpacesStrippedCommand = cmd.replace(/\s{2,}/g, ' ')
+  const splitCommand = extraSpacesStrippedCommand.split(/\s+(?![^[]*]|[^<]*>)/)
+  const bregex = /\.*[\][<>]/g
+
+  const parsedCommand: ParsedCommand = {
+    cmd: (splitCommand.shift()).replace(bregex, ''),
+    demanded: [],
+    optional: []
+  }
+  splitCommand.forEach((cmd, i) => {
+    let variadic = false
+    cmd = cmd.replace(/\s/g, '')
+    if (/\.+[\]>]/.test(cmd) && i === splitCommand.length - 1) variadic = true
+    if (/^\[/.test(cmd)) {
+      parsedCommand.optional.push({
+        cmd: cmd.replace(bregex, '').split('|'),
+        variadic
+      })
+    } else {
+      parsedCommand.demanded.push({
+        cmd: cmd.replace(bregex, '').split('|'),
+        variadic
+      })
+    }
+  })
+  return parsedCommand
+}
