+++ lib/types/positional.ts
@@ -1,0 +1,4 @@
+export interface Positional {
+  cmd: string[]
+  variadic: boolean
+}
